chore: minor cleanup of some schema definitions (#1723)

+++ packages/common-utils/src/types.ts
@@ -374,7 +374,7 @@ export type SavedSearch = z.infer<typeof SavedSearchSchema>;
 // --------------------------
 export const NumberFormatSchema = z.object({
   output: z.enum(['currency', 'percent', 'byte', 'time', 'number']),
-  mantissa: z.number().optional(),
+  mantissa: z.number().int().optional(),
   thousandSeparated: z.boolean().optional(),
   average: z.boolean().optional(),
   decimalBytes: z.boolean().optional(),
